requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
bleach==4.1.0
boto3==1.26.51
botocore==1.29.51
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-address==0.2.8
django-admin-relation-links==0.2.5
django-clearcache==1.2.1
django-cors-headers==3.13.0
django-redis==5.2.0
django-richtextfield==1.6.1
django-storages==1.13.2
django-summernote==0.8.20.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
dnspython==2.3.0
et-xmlfile==1.1.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
numpy==1.24.1
oauthlib==3.2.2
openpyxl==3.0.10
packaging==23.0
pandas==1.5.3
pip-chill==1.0.1
psycogreen==1.0.2
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pymemcache==4.0.0
pymongo==4.3.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
redis==4.5.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
supervisor==4.2.5
tinycss2==1.1.1
uritemplate==4.1.1
urllib3==1.26.13
webencodings==0.5.1
zope.event==4.6
zope.interface==5.5.2
